Refactored the Recharts integration to lazy-load all chart components dynamically, preventing them from being bundled into the shared pages bundle.
